refactor(native): move keyboard tests next to the code they cover (#2831)
Distribute the per-domain test clusters into their modules: controller-mapping
tests (button/port routing, snes_key_to_button_id) -> controller_mapping.rs,
hotkey/cart-switch tests -> hotkeys.rs, per-console dispatch tests ->
console_keyboard.rs; keyboard_target_ports tests stay in mod.rs. All share the
test_support helpers. No behavior change (88 tests pass).
